refactor: move Saml logic to Sso namespace
