+++ EMS/client-helper-bundle/src/Security/Sso/Saml/SamlAuthFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EMS\ClientHelperBundle\Security\Saml;
+namespace EMS\ClientHelperBundle\Security\Sso\Saml;
 
 use OneLogin\Saml2\Auth;
 use OneLogin\Saml2\Utils;
