+++ elasticms-web/config/packages/security.yaml
@@ -5,7 +5,7 @@ security:
         core_api:
             id: emsch.security.core_api.user_provider
         saml:
-            id: emsch.security.saml.user_provider
+            id: emsch.security.sso.saml.user_provider
         chain_providers:
             chain:
                 providers: [ 'core_api', 'saml' ]
@@ -20,7 +20,7 @@ security:
             provider: chain_providers
             custom_authenticators:
                 - emsch.security.core_api.authenticator
-                - emsch.security.saml.authenticator
+                - emsch.security.sso.saml.authenticator
             logout:
                 path: emsch_logout
